+++ src/main/java/com/cookingchef/controller/RecipeController.java
@@ -3,37 +3,24 @@ package com.cookingchef.controller;
 import com.cookingchef.facade.CategoryFacade;
 import com.cookingchef.facade.IngredientFacade;
 import com.cookingchef.facade.RecipeFacade;
-import com.cookingchef.model.Category;
-import com.cookingchef.model.CategoryRecipe;
-import com.cookingchef.model.IngredientRecipe;
-import com.cookingchef.model.Recipe;
-import com.cookingchef.view.Main;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
+import com.cookingchef.model.*;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
-import javafx.scene.Scene;
-import javafx.scene.control.*;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.control.ListCell;
+import javafx.scene.control.ListView;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
-import javafx.scene.image.PixelReader;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
-import javafx.stage.FileChooser;
-import javafx.stage.Modality;
-import javafx.stage.Stage;
-import org.controlsfx.control.CheckComboBox;
-import org.controlsfx.control.Notifications;
+import javafx.scene.text.Text;
 
-import javax.imageio.ImageIO;
-import java.awt.image.BufferedImage;
+import javax.swing.text.View;
 import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.IOException;
 import java.net.URL;
-import java.nio.file.Files;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
@@ -41,395 +28,110 @@ import java.util.ResourceBundle;
 
 public class RecipeController implements Initializable {
     @FXML
-    private TextField name;
-
-    @FXML
-    private TextArea description;
-
-    @FXML
-    private TextField summary;
-
-    @FXML
     private ImageView imageView;
 
     @FXML
-    private TextField servings;
+    private Label recipeName;
 
     @FXML
-    private CheckComboBox<IngredientRecipe> ingredientRecipeCheckComboBox;
+    private Text recipeDescription;
 
     @FXML
-    private CheckComboBox<CategoryRecipe> categoryRecipeCheckComboBox;
+    private Text recipeSummary;
 
     @FXML
-    private Button cancelButton;
+    private Text recipeServings;
 
     @FXML
-    private Button validateButton;
+    private ListView<Ingredient> ingredientListView = new ListView<Ingredient>();
 
     @FXML
-    private Button addImage;
+    private ListView<Category> categoryListView = new ListView<Category>();
 
     @FXML
-    private Stage secondaryStage;
-
-    @FXML
-    private ListView<Recipe> recipeListView = new ListView<Recipe>();
-
-    private final RecipeFacade recipeFacade;
+    private Button retourButton;
     private final IngredientFacade ingredientFacade;
     private final CategoryFacade categoryFacade;
 
-    public RecipeController(){
-        this.recipeFacade = RecipeFacade.getRecipeFacade();
+    private final Recipe recipe;
+
+    public RecipeController(Recipe recipe){
+        this.recipe = recipe;
         this.ingredientFacade = IngredientFacade.getIngredientFacade();
         this.categoryFacade = CategoryFacade.getCategoryFacade();
     }
 
-    public List<Recipe> showList(){
-        List<Recipe> recipes;
-
-        try {
-            recipes = this.recipeFacade.getAllRecipes();
-        } catch (SQLException e) {
-            e.printStackTrace();
-            throw new RuntimeException(e);
-        }
-        if (recipes == null) {
-            Notifications.create()
-                    .title("Erreur")
-                    .text("Erreur avec la base de donnée\nRelancer l'application")
-                    .showError();
-        }
-        return recipes;
-    }
-
-    public void deleteRecipe(Recipe recipe) {
-        try {
-            this.recipeFacade.deleteRecipe(recipe);
-
-            this.recipeListView.getItems().remove(recipe);
-            Notifications.create()
-                    .title("Succès")
-                    .text("La recette a été supprimée avec succès")
-                    .showConfirm();
-        } catch (SQLException e) {
-            Notifications.create()
-                    .title("Erreur")
-                    .text("Erreur lors de la suppression")
-                    .showError();
-            e.printStackTrace();
-            throw new RuntimeException(e);
-        }
-    }
-
-    public void validateForm(){
-        if (this.name.getText().isEmpty()) {
-            Notifications.create()
-                    .title("Information")
-                    .text("Veuillez remplir le nom de la recette")
-                    .showWarning();
-
-        }
-
-        if (this.description.getText().isEmpty()) {
-            Notifications.create()
-                    .title("Information")
-                    .text("Veuillez remplir la description de la recette")
-                    .showWarning();
-
-        }
-
-        if (this.summary.getText().isEmpty()) {
-            Notifications.create()
-                    .title("Information")
-                    .text("Veuillez remplir le résumé de la recette")
-                    .showWarning();
-
-        }
-
-        if (this.servings.getText().isEmpty()) {
-            Notifications.create()
-                    .title("Information")
-                    .text("Veuillez remplir le nombre de personnes pour la recette")
-                    .showWarning();
-
-        }
-    }
-
-    private byte[] addImageRecipe(Stage stage) {
-        FileChooser fileChooser = new FileChooser();
-        File file = fileChooser.showOpenDialog(stage);
-        if (file != null) {
-            try {
-                return Files.readAllBytes(file.toPath());
-            } catch (IOException e) {
-                // TODO : gérer exception
-                e.printStackTrace();
-            }
-        }
-        return null;
-    }
-
-    private byte[] recupImage() {
-        // Récupérer l'image séléctionnée
-        PixelReader pixelReader = imageView.getImage().getPixelReader();
-        int width = (int) imageView.getImage().getWidth();
-        int height = (int) imageView.getImage().getHeight();
-
-        BufferedImage bImage = new BufferedImage(width, height, BufferedImage.TYPE_INT_ARGB);
-        for (int y = 0; y < height; y++) {
-            for (int x = 0; x < width; x++) {
-                bImage.setRGB(x, y, pixelReader.getArgb(x, y));
-            }
-        }
-
-        byte[] imageData = new byte[0];
-        try {
-            File tempFile = File.createTempFile("image", ".tmp");
-            ImageIO.write(bImage, "png", tempFile);
-
-            imageData = Files.readAllBytes(tempFile.toPath());
-            tempFile.delete();
-        } catch (IOException e) {
-            // TODO : gérer l'exception
-            e.printStackTrace();
-        }
-
-        return imageData;
-    }
-
-    public void showFormCreate() {
-        // Afficher le formulaire général de create Ingredient
-        this.showForm();
-
-        // Définir le bouton valider pour le create form
-        this.validateButton.setOnAction(actionEvent -> {
-
-            this.validateForm();
-
-            Recipe recipe = new Recipe(
-                    this.name.getText(),
-                    this.description.getText(),
-                    this.summary.getText(),
-                    this.recupImage(),
-                    Integer.parseInt(this.servings.getText()),
-                    new ArrayList<IngredientRecipe>(this.ingredientRecipeCheckComboBox.getCheckModel().getCheckedItems()),
-                    new ArrayList<CategoryRecipe>(this.categoryRecipeCheckComboBox.getCheckModel().getCheckedItems())
-            );
-
-            this.createRecipe(recipe);
-        });
-    }
-
-    private void createRecipe(Recipe recipe) {
-        try {
-            this.recipeFacade.addRecipe(recipe);
-            this.secondaryStage.close();
-            Notifications.create()
-                    .title("Succès")
-                    .text("La recette a été créé avec succès")
-                    .showConfirm();
-            this.showList();
-        } catch (SQLException e) {
-            Notifications.create()
-                    .title("Erreur")
-                    .text("Erreur lors de la création")
-                    .showError();
-            e.printStackTrace();
-            throw new RuntimeException(e);
-        }
-    }
-
-    public void showFormUpdate(Recipe recipe) {
-        this.showForm();
-
-        // Remplir avec les informations existantes
-        this.name.setText(recipe.getName());
-        this.description.setText(recipe.getDescription());
-        this.summary.setText(recipe.getSummary());
-        this.imageView.setImage(new Image(new ByteArrayInputStream(recipe.getSrc())));
-        this.servings.setText(String.valueOf(recipe.getServings()));
-        // TODO : remplir les listes de catégories et d'ingrédients
-        this.ingredientRecipeCheckComboBox.getCheckModel().checkIndices(recipe.getListOfIngredients());
-        this.categoryRecipeCheckComboBox.getCheckModel().checkIndices(recipe.getListOfCategories());
-
-        // Définir le bouton valider pour le update form
-        this.validateButton.setOnAction(actionEvent -> {
-
-            this.validateForm();
-
-            Recipe recipeToUpdate = new Recipe(
-                    this.name.getText(),
-                    this.description.getText(),
-                    this.summary.getText(),
-                    this.recupImage(),
-                    Integer.parseInt(this.servings.getText()),
-                    new ArrayList<IngredientRecipe>(this.ingredientRecipeCheckComboBox.getCheckModel().getCheckedItems()),
-                    new ArrayList<CategoryRecipe>(this.categoryRecipeCheckComboBox.getCheckModel().getCheckedItems())
-            );
-
-            this.updateCategory(recipeToUpdate);
-        });
-    }
-
-    private void updateCategory(Recipe recipeToUpdate) {
-        try {
-            this.recipeFacade.updateRecipe(recipeToUpdate);
-            this.secondaryStage.close();
-            Notifications.create()
-                    .title("Sucess")
-                    .text("La catégorie a été mise à jour avec succès")
-                    .showConfirm();
-            this.showList();
-        } catch (SQLException e) {
-            Notifications.create()
-                    .title("Erreur")
-                    .text("Erreur lors de la modification")
-                    .showError();
-            e.printStackTrace();
-            throw new RuntimeException(e);
-        }
-    }
-
-    private void showForm() {
-        // Création de la seconde fenêtre
-        this.secondaryStage = new Stage();
-        secondaryStage.initModality(Modality.WINDOW_MODAL);
-        secondaryStage.initOwner(Main.stage);
-
-        Label labelName = new Label("Nom de la recette");
-        this.name = new TextField();
-
-        Label labelDescription = new Label("Description de la recette");
-        this.description = new TextArea();
-
-        Label labelSummary = new Label("Résumé de la recette");
-        this.summary = new TextField();
-
-        Label labelImage = new Label("Image de la recette");
-        this.imageView = new ImageView();
-        this.imageView.setFitHeight(150);
-        this.imageView.setFitWidth(188.0);
-        this.imageView.setPickOnBounds(true);
-        this.imageView.setPreserveRatio(true);
-        this.addImage = new Button("Importer image");
-        addImage.setOnAction(actionEvent -> {
-            byte[] bufferImage = this.addImageRecipe(this.secondaryStage);
-            if (bufferImage != null) {
-                this.imageView.setImage(new Image(new ByteArrayInputStream(bufferImage)));
-            } else {
-                Notifications.create()
-                        .title("Information")
-                        .text("Image inchangée")
-                        .showInformation();
-            }
-        });
-
-        Label labelServings = new Label("Nombre de personnes");
-        this.servings = new TextField();
-
-        Label labelIngredient = new Label("Ingrédients");
-        this.ingredientRecipeCheckComboBox = new CheckComboBox<IngredientRecipe>();
-
-        Label labelCategory = new Label("Catégories");
-        this.categoryRecipeCheckComboBox = new CheckComboBox<CategoryRecipe>();
-
-        this.cancelButton = new Button("Annuler");
-        this.cancelButton.setOnAction(actionEvent -> secondaryStage.close());
-
-        this.validateButton = new Button("Valider");
-
-        HBox first = new HBox(labelName, this.name);
-        first.setSpacing(10);
-        first.setPadding(new Insets(10));
-        first.setAlignment(Pos.CENTER);
-
-        HBox second = new HBox(labelSummary, this.summary);
-        second.setSpacing(10);
-        second.setPadding(new Insets(10));
-        second.setAlignment(Pos.CENTER);
-
-        HBox third = new HBox(labelServings, this.servings);
-        third.setSpacing(10);
-        third.setPadding(new Insets(10));
-        third.setAlignment(Pos.CENTER);
-
-        HBox fourth = new HBox(labelCategory, this.categoryRecipeCheckComboBox);
-        fourth.setSpacing(10);
-        fourth.setPadding(new Insets(10));
-        fourth.setAlignment(Pos.CENTER);
-
-        HBox five = new HBox(labelIngredient, this.ingredientRecipeCheckComboBox);
-        five.setSpacing(10);
-        five.setPadding(new Insets(10));
-        five.setAlignment(Pos.CENTER);
-
-        HBox six = new HBox(labelDescription, this.description);
-        six.setSpacing(10);
-        six.setPadding(new Insets(10));
-        six.setAlignment(Pos.CENTER);
-
-        HBox seven = new HBox(labelImage, this.imageView, this.addImage);
-        seven.setSpacing(10);
-        seven.setPadding(new Insets(10));
-        seven.setAlignment(Pos.CENTER);
-
-        HBox eight = new HBox(this.cancelButton, this.validateButton);
-        eight.setSpacing(10);
-        eight.setPadding(new Insets(10));
-        eight.setAlignment(Pos.CENTER);
-
-        VBox root = new VBox(first, second, third, fourth, five, six, seven, eight);
-        root.setSpacing(10);
-        root.setPadding(new Insets(10));
-        root.setAlignment(Pos.CENTER);
-        Scene scene = new Scene(root, 500, 400);
-        this.secondaryStage.setScene(scene);
-        this.secondaryStage.show();
-    }
-
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
-        ObservableList<Recipe> recipes = FXCollections.observableArrayList(this.showList());
+        this.recipeName = new Label(recipe.getName());
+        this.recipeDescription = new Text(recipe.getDescription());
+        this.recipeSummary = new Text(recipe.getSummary());
+        this.recipeServings = new Text(String.valueOf(recipe.getServings()));
 
-        recipeListView.setItems(recipes);
 
-        recipeListView.setCellFactory(param -> new ListCell<>() {
+        List<IngredientRecipe> ingredientsId = this.recipe.getListOfIngredients();
+        List<Ingredient> ingredients = new ArrayList<Ingredient>();
+        for (IngredientRecipe ingredientId : ingredientsId) {
+            try {
+                ingredients.add(this.ingredientFacade.getIngredientById(ingredientId.getIngredient()));
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
+            }
+        }
+        this.ingredientListView.getItems().addAll(ingredients);
+
+        List<CategoryRecipe> categoryRecipes = this.recipe.getListofCategories();
+        List<Category> categories = new ArrayList<Category>();
+        for (CategoryRecipe categoryRecipe : categoryRecipes) {
+            try {
+                categories.add(this.categoryFacade.getCategoryRecipeById(categoryRecipe.getCategoryId()));
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
+            }
+        }
+        this.categoryListView.getItems().addAll(categories);
+
+        this.imageView = new ImageView(new Image(new ByteArrayInputStream(recipe.getSrc())));
+
+        categoryListView.setCellFactory(param -> this.listCellFactoryCategory());
+        ingredientListView.setCellFactory(param -> this.listCellFactoryIngredient());
+    }
+
+    public ListCell<Category> listCellFactoryCategory() {
+        return new ListCell<Category>() {
 
             @Override
-            protected void updateItem(Recipe recipe) {
-                super.updateItem(recipe);
+            protected void updateItem(Category category, boolean empty) {
+                super.updateItem(category, empty);
 
-                if (empty || category == null || category.getValue().getNameCategory() == null) {
+                if (empty || category == null || category.getNameCategory() == null) {
                     setText(null);
                 } else {
                     VBox vBox = new VBox();
-                    Label label = new Label(category.getKey() + " : " + category.getValue().getNameCategory());
-
-                    Button deleteButton = new Button("Supprimer");
-                    deleteButton.setOnAction(actionEvent -> {
-                        Category categoryToDelete = getItem().getValue();
-                        // code de suppression de l'ingrédient de la ListView
-                        deleteCategory(getItem().getKey(), categoryToDelete.getIdCategory());
-                    });
-
-                    Button updateButton = new Button("Modifer");
-                    updateButton.setOnAction(actionEvent -> {
-                        Category categoryToUpdate = getItem().getValue();
-                        // code de modification de l'ingrédient de la ListView
-                        showFormUpdate(getItem().getKey(), categoryToUpdate);
-                    });
-
-                    HBox hBox = new HBox(deleteButton, updateButton);
-                    hBox.setSpacing(10);
-                    hBox.setPadding(new Insets(10));
-                    hBox.setAlignment(Pos.CENTER);
-
-                    vBox.getChildren().addAll(label, hBox);
+                    Label nameLabel = new Label(category.getNameCategory());
+                    vBox.getChildren().addAll(nameLabel);
                     setGraphic(vBox);
                 }
             }
-        });
+        };
+    }
+
+    public ListCell<Ingredient> listCellFactoryIngredient() {
+        return new ListCell<Ingredient>() {
+
+            @Override
+            protected void updateItem(Ingredient ingredient, boolean empty) {
+                super.updateItem(ingredient, empty);
+
+                if (empty || ingredient == null || ingredient.getName() == null) {
+                    setText(null);
+                } else {
+                    VBox vBox = new VBox();
+                    Label nameLabel = new Label(ingredient.getName());
+                    vBox.getChildren().addAll(nameLabel);
+                    setGraphic(vBox);
+                }
+            }
+        };
     }
 }
