+++ src/main/resources/com/cookingchef/view/recipes/recipe-admin-view.fxml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.effect.InnerShadow?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<BorderPane prefHeight="1080.0" prefWidth="1920.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.cookingchef.controller.RecipeAdminController">
+    <center>
+        <VBox>
+         <HBox prefHeight="100.0" prefWidth="200.0">
+            <children>
+                    <Label alignment="CENTER" contentDisplay="CENTER" prefHeight="76.0" prefWidth="346.0" text="Liste des recettes :" underline="true">
+                        <font>
+                            <Font size="28.0" />
+                        </font>
+                        <effect>
+                            <InnerShadow />
+                        </effect>
+                    </Label>
+            </children>
+         </HBox>
+         <HBox prefHeight="30.0" prefWidth="1860.0">
+            <children>
+                   <Button contentDisplay="CENTER" mnemonicParsing="false" onAction="#showFormCreate" prefWidth="151.0" text="Ajouter une recette">
+                  <font>
+                     <Font size="14.0" />
+                  </font>
+                  <HBox.margin>
+                     <Insets />
+                  </HBox.margin>
+               </Button>
+            </children>
+            <VBox.margin>
+               <Insets left="60.0" />
+            </VBox.margin>
+         </HBox>
+            <ScrollPane>
+                <content>
+                    <ListView fx:id="recipeListView" prefHeight="400.0" prefWidth="1757.0" />
+                </content>
+                <VBox.margin>
+                    <Insets left="80.0" right="80.0" top="40.0" />
+                </VBox.margin>
+            </ScrollPane>
+        </VBox>
+
+    </center>
+</BorderPane>
